Renamed files, cleaned them up, fixed the expiration time.
