+++ backend/.env-sample
@@ -1,1 +1,3 @@
 PORT=
+ENVIRONMENT=
+LINK_EXPIRATION_MINUTES=
