+++ frontend/src/App.jsx
@@ -1,17 +1,11 @@
-import { useState } from "react";
 import { Routes, Route } from "react-router";
 import FileUpload from "./components/FileUpload";
-import GrokFileUpload from "./components/GrokFileUpload";
-import GrokFileUploadModified from "./components/GrokFileUploadModified";
 import ViewImage from "./pages/ViewImage";
-import TutorialFileUpload from "./components/TutorialFileUpload";
 import Grid from "@mui/material/Grid";
 import CssBaseline from "@mui/material/CssBaseline";
 import NotFound from "./pages/NotFound";
 
 function App() {
-  const [count, setCount] = useState(0);
-
   return (
     <>
       <CssBaseline />
@@ -25,15 +19,12 @@ background: linear-gradient(30deg, rgba(13, 101, 184, 1) 11%, rgba(17, 131, 237,
       >
         <Routes>
           <Route path="/notfound" element={<NotFound />} />
-          <Route path="/" element={<GrokFileUploadModified />} />
+          <Route path="/" element={<FileUpload />} />
           <Route
             path="/viewImage/:fileName"
             element={<ViewImage />}
           />
         </Routes>
-        {/* <TutorialFileUpload /> */}
-        {/* <GrokFileUpload/> */}
-        {/* <FileUpload /> */}
       </Grid>
     </>
   );
